fix warnings, make sure all doubles and no floats, make sure equal override in Vector2 is const

+++ TFMatcher/include/SharedDefinitions.h
@@ -1,6 +1,6 @@
 #ifndef _SHARED_DEFINITIONS_H_
 #define _SHARED_DEFINITIONS_H_
 
-constexpr float EPSILON = 1e-10;
+constexpr double EPSILON = 1e-10;
 
 #endif
